I reworked the templates to incorporate the new logo. Some pages have some slightly awkward margins and such, but I'll will clean them up.
The front page also looks a little funky with the blue tongue and the new blue nav bar. I'm going to look into dumping the tongue and generally reworking the front page to bring it up to spec with the rest of the site. 

No php changes, just CSS and templates.